chore: change api call to get categories to hook
